requirements.txt
# Azurebrains RRSS — dependencias Python
# Actualizar versiones con: pip-compile requirements.in

# HTTP
httpx>=0.27.0

# Azure SDK
azure-cosmos>=4.9.0
azure-identity>=1.17.0

# Azure OpenAI
openai>=1.30.0

# X (Twitter) API v2
tweepy>=4.14.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.7.0
